brew_hub/requirements.txt
Django==5.0.4
gunicorn==20.1.0
psycopg2-binary==2.9.9
nanoid==2.0.0
django-allauth~=0.63.6
rollbar~=1.0
django-cotton==1.5.1
django-bulma==0.9.0
